New Controller and Control to Vrep Converter Nodes
Movement topics are now generated using pwm values and are incrementally
maxed in at least 2 seconds (depending on joystick y-axis and trigger
values). The converter node simply takes these values and converts them
to target speeds for the simulator. Bin actuation in the controller was
also added, but only using a position value (which is at least supported
by the simulation).
On a minor note, the simulation environment was adjusted to be a 25x25ft
arena.
A problem was noticed with the simulated model being unable to handle
the full load of dirt. When the joints were moved into a more vertical
position and their target ranges extended, a jittering anomaly occurred
on the right piston; the changes were undone before this commit.
Remade the workspace (which explains the ton of commits). The build file
was removed but the devel file was left as it includes generated files
needed by v-rep (which contains references to them, at least in my
machine).